refactor: moved budget performance to analysis package
